trust/requirements.txt
Django==2.2.16
dnspython=2.4.2
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.19.0